BinarySearch implemented

+++ app/src/main/java/com/github/cythara/PitchDifference.java
@@ -26,7 +26,7 @@ class PitchDifference implements Parcelable {
     private PitchDifference(Parcel in) {
         Tuning tuning = MainActivity.getCurrentTuning();
         closest = tuning.findNote(in.readString());
-        deviation = in.readDouble();
+        deviation = in.readFloat();
     }
 
     @Override
